move resolvers to cli/jsr.rs
